fix logging in isr
